+++ projects/ngx-testing-tools/src/lib/directive-test-bed/directive-test-bed-factory.ts
@@ -2,8 +2,8 @@ import { ElementRef, Type, ViewContainerRef } from '@angular/core';
 import { assertDirectiveCtor } from '../common/assertions/assert-directive-ctor';
 import { RendererTestBedFactory } from '../common/test-beds/renderer/renderer-test-bed-factory';
 import { InjectionStore } from '../common/tools/store/models/injected-store.model';
-import { HOST_FIXTURE } from './host/host-component';
 import { HostElementRef } from './host/host-element-ref';
+import { HOST_FIXTURE } from './host/host-fixture.token';
 import { DirectiveTestBedOptions } from './models';
 import { DirectiveTools } from './tools';
 import { buildDirectiveTools } from './tools/directive-tools';
