+++ projects/ngx-testing-tools/src/lib/interceptor-test-bed/interceptor-test-bed-factory.ts
@@ -7,7 +7,7 @@ import { httpProviders } from '../common/tools/http/http-providers';
 import { InjectionStore } from '../common/tools/store/models/injected-store.model';
 import { isConstructor } from '../common/utils/constructor.util';
 import { InterceptorTestBedOptions } from './models';
-import { INTERCEPTOR_INFO, InterceptorInfo } from './proxy/interceptor-info';
+import { INTERCEPTOR_INFO, InterceptorInfo } from './proxy/interceptor-info.token';
 import { InterceptorProxy } from './proxy/interceptor-proxy';
 import { InterceptorTools } from './tools';
 import { buildInterceptorTools } from './tools/interceptor-tools';
